requirements.txt
PySide6>=6.6
requests
fpdf2
opencv-python
numpy
python-dotenv
